% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-visualization.R
\name{as_cadenza_visualization}
\alias{as_cadenza_visualization}
\title{Prepare visualization results for Cadenza}
\usage{
as_cadenza_visualization(res, plotfile)
}
\arguments{
\item{res}{The Plumber response construct}

\item{plotfile}{The plot file created in the Plumber R file}
}
\value{
The Plumber response construct with the plot data added
}
\description{
Prepare visualization results for Cadenza
}
\examples{
png(file = "plot.png")
plot(list(1), list(1))
dev.off()
res <- list("injected from plumber")
res <- as_cadenza_visualization(res, "plot.png")
res$status
res$headers
length(res$body)
}
