% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-capabilities.R
\name{attribute_group}
\alias{attribute_group}
\title{A group of attributes used in the extension.}
\usage{
attribute_group(...)
}
\arguments{
\item{...}{key value parameters, requires name and printname.}
}
\description{
See the examples in \code{\link[=serializer_cadenza_capabilities]{serializer_cadenza_capabilities()}}.
}
