% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-parsers.R
\name{detect_boundary}
\alias{detect_boundary}
\title{Find the boundary in the content type multipart MIME header}
\usage{
detect_boundary(content_type)
}
\arguments{
\item{content_type}{the content of the content_type mime header.}
}
\description{
Find the boundary in the content type multipart MIME header
}
