% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\docType{data}
\name{empty_named_list}
\alias{empty_named_list}
\title{Empty Named List}
\format{
An object of class \code{list} of length 0.
}
\usage{
empty_named_list
}
\description{
When building a JSON from an R list of lists, if you want to have
an empty object instead of an empty array, you can use this object.
}
\examples{
# a is parsed as an empty object, while b is parsed as an empty array
jsonlite::toJSON(list(a = empty_named_list, b = list()))
# returns {"a":{},"b":[]}
}
\keyword{datasets}
