% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-capabilities.R
\name{extension}
\alias{extension}
\title{The extension description}
\usage{
extension(...)
}
\arguments{
\item{...}{parameters of the analytics extension. It requires a printName, extensionType, and attributeGroups.}
}
\description{
See the examples in \code{\link[=serializer_cadenza_capabilities]{serializer_cadenza_capabilities()}}.
}
