% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-discovery.R
\name{extension_reference}
\alias{extension_reference}
\title{The description of one extension}
\usage{
extension_reference(...)
}
\arguments{
\item{...}{requires a printName, extensionType, and attributeGroups.
It can optionally take additional parameters.
For examples, see \code{\link[=serializer_cadenza_discovery]{serializer_cadenza_discovery()}}}
}
\description{
The description of one extension
}
