% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-parsers.R
\name{parser_cadenza}
\alias{parser_cadenza}
\alias{parser_delim}
\title{Custom parsers}
\usage{
parser_cadenza()

parser_delim(...)
}
\arguments{
\item{...}{Additional arguments passed on to \code{\link[vroom:vroom]{vroom::vroom()}}.}
}
\value{
A list with the following elements:
\itemize{
\item data: the data sent by Cadenza; Can be a dataframe or a list
\item column_info: a dataframe with metadata regarding the columns in data if it is a dataframe
\item metadata: a list with all metadata supplied by Cadenza
\item ... Further elements linked to parameters specified in the extension description
}
}
\description{
A parser is responsible for returning a valid R object based on the document-type.
The Cadenza parser takes a multipart form with a specific structure and returns a dataframe.
}
\details{
See \code{\link[plumber:parsers]{plumber::parser_form()}} for details and
\code{\link[plumber:register_parser]{plumber::registered_parsers()}} for a list of registered parsers names.
}
\section{Functions}{
\itemize{
\item \code{parser_cadenza()}: Parse multipart-form request from Cadenza

\item \code{parser_delim()}: Parse any delimited files. See \code{\link[vroom:vroom]{vroom::vroom()}} for details.

}}
