% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plumbercadenza-package.R
\docType{package}
\name{plumbercadenza-package}
\alias{plumbercadenza}
\alias{plumbercadenza-package}
\title{plumbercadenza: Create An API For The Cadenza Analytics Extension With Plumber}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Matthias Theobald \email{matthias.theobald@disy.net}

Authors:
\itemize{
  \item Lossow Stefan \email{stefan.lossow@disy.net}
}

Other contributors:
\itemize{
  \item Disy [copyright holder]
}

}
\keyword{internal}
