% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-capabilities.R
\name{register_serializer_capabilities_onLoad}
\alias{register_serializer_capabilities_onLoad}
\title{register the defined serializer}
\usage{
register_serializer_capabilities_onLoad()
}
\description{
register the defined serializer
}
