% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-enrichment-calculation.R,
%   R/custom-serializer-text.R, R/custom-serializer-url.R,
%   R/custom-serializer-visualization.R
\name{serializer_multi}
\alias{serializer_multi}
\alias{serializer_cadenza_enrichment_calculation}
\alias{serializer_cadenza_text}
\alias{serializer_cadenza_url}
\alias{serializer_cadenza_visualization}
\title{Custom serializers}
\usage{
serializer_multi(serialize_fn, boundary = NULL, ...)

serializer_cadenza_enrichment_calculation(...)

serializer_cadenza_text(...)

serializer_cadenza_url(...)

serializer_cadenza_visualization(...)
}
\arguments{
\item{serialize_fn}{Function with which the individual parts will
be serialized.  The function has to return a vector or list of
strings and has to have the attribute \code{type} with the HTML
content type of each part.}

\item{boundary}{The boundary string with which the parts are to be
separated.  If \code{NULL}, a random string of 30 characters will be
generated and used.}

\item{...}{Additional arguments to be passed to the serialize
function.}
}
\description{
Custom serializers
}
\section{Functions}{
\itemize{
\item \code{serializer_multi()}: Serialize as multipart-form

\item \code{serializer_cadenza_enrichment_calculation()}: Serialize the multipart-form required by Cadenza for
enrichments and calculations

\item \code{serializer_cadenza_text()}: Serialize the multipart-form required by
Cadenza for text

\item \code{serializer_cadenza_url()}: Serialize the multipart-form required by
Cadenza for text

\item \code{serializer_cadenza_visualization()}: Serialize the multipart-form required by
Cadenza for visualizations

}}
\examples{
\dontrun{
pr() |>
  pr_get(
    "/form",
    handler = function() list(data.frame(x = rnorm(10), y = runif(10)),
                              list(param = "b", value = 12)),
    serializer = serializer_multi(
      serialiize_fn = function(val) lapply(val, )))
}
}
