% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-visualization.R
\name{as_cadenza_visualization}
\alias{as_cadenza_visualization}
\title{Prepare visualization results for Cadenza}
\usage{
as_cadenza_visualization(plotfile)
}
\arguments{
\item{plotfile}{The plot file created in the Plumber R file}
}
\value{
res The Plumber reponse construct
}
\description{
Prepare visualization results for Cadenza
}
\examples{
res <- as_cadenza_visualization(plotfile)
}
