% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-discovery.R
\name{discovery}
\alias{discovery}
\title{The description for available extensions}
\usage{
discovery(...)
}
\arguments{
\item{extensions}{a list of \code{\link[=extension_reference]{extension_reference()}}'s. For examples, see \code{\link[=serializer_cadenza_discovery]{serializer_cadenza_discovery()}}}
}
\description{
The description for available extensions
}
