% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-capabilities.R
\name{parameter}
\alias{parameter}
\title{A parameter that can or must be passed to the extension}
\usage{
parameter(...)
}
\description{
A parameter that can or must be passed to the extension
}
