% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-capabilities.R
\name{serializer_cadenza_capabilities}
\alias{serializer_cadenza_capabilities}
\title{Build the Capabilities GET-response.}
\usage{
serializer_cadenza_capabilities(...)
}
\arguments{
\item{extension}{A list of a parameters of the analytics extension.
It requires a printName, extensionType, and attributeGroups.
It can optionally take parameters.}
}
\value{
a GetCapabilities response.
}
\description{
Build the Capabilities GET-response.
}
\examples{
function() {
  extension(
    printName = "Data Generation",
    extensionType = "calculation",
    attributeGroups = list(
      attribute_group(
        name = "datgen",
        printName = "Data Generation",
        dataTypes = c("int64", "float64"),
        maxAttributes=1L
      )
    ),
    parameters = list(
      parameter(
        name = "cols",
        printName = "Please select the number of columns to generate.",
        parameterType = "select",
        options = c(2, 3, 4, 5, 6, 7, 8, 9, 10),
        required = TRUE,
        defaultValue = c("2")
      )
    )
  )
}
}
