% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom-serializer-discovery.R
\name{serializer_cadenza_discovery}
\alias{serializer_cadenza_discovery}
\title{Build the Discovery GET-response.}
\usage{
serializer_cadenza_discovery(...)
}
\arguments{
\item{discovery}{A list with an extension of a extension_references.
Each requires a printName, extensionType, and attributeGroups.
It can optionally take parameters.}
}
\value{
a discovery response.
}
\description{
Build the Discovery GET-response.
}
\examples{
function() {
  list(
    extension_reference(
      extensionPrintName = "The Name",
      extensionType = "calculation",
      relativePath = "/path-of-extension"
    )
  )
}
}
